refactor(undeploy): split the plan vocabulary out of runUndeploy
